26.03.24

See merge request s14-bigdata-recom-sub1/S14P21A505!153

+++ back/src/main/java/com/a505/jupasu/domain/reviews/dto/ReviewResponse.java
@@ -13,7 +13,7 @@ public class ReviewResponse {
     private Long reviewId;
     private Long userId;
     private String nickname;
-    private Float rating;
+    private Integer rating;
     private String content;
     private LocalDateTime createdAt;
